+++ Snyk.VisualStudio.Extension.2022/Settings/IPersistableOption.cs
@@ -37,6 +37,7 @@ public interface IPersistableOptions
     /// </summary>
     string Organization { get; set; }
 
+
     /// <summary>
     /// Gets or sets a value indicating whether CLI ignore unknown CA parameter.
     /// </summary>