+++ Snyk.VisualStudio.Extension.2022/Settings/SnykScanOptionsUserControl.cs
@@ -16,11 +16,6 @@ namespace Snyk.VisualStudio.Extension.Settings
         private static readonly ILogger Logger = LogManager.ForContext<SnykScanOptionsUserControl>();
         public ISnykOptions OptionsMemento { get; set; }
 
-        private static readonly int TwoSecondsDelay = 2000;
-
-        private const int MaxSastRequestAttempts = 20;
-
-        private readonly Timer snykCodeEnableTimer = new Timer();
 
         public SnykScanOptionsUserControl(ISnykServiceProvider serviceProvider)
         {
@@ -43,94 +38,14 @@ namespace Snyk.VisualStudio.Extension.Settings
 
         private void SnykScanOptionsUserControl_Load(object sender, EventArgs e)
         {
-            this.StartSastEnablementCheckLoop();
-        }
-
-        private void StartSastEnablementCheckLoop()
-        {
-            try
-            {
-                if (this.snykCodeEnableTimer.Enabled)
-                {
-                    this.snykCodeEnableTimer.Stop();
-                }
-
-                var currentRequestAttempt = 1;
-
-                this.snykCodeEnableTimer.Interval = TwoSecondsDelay;
-
-                this.snykCodeEnableTimer.Tick += (sender, args) => ThreadHelper.JoinableTaskFactory.RunAsync(async () =>
-                {
-                    try
-                    {
-                        if (!LanguageClientHelper.IsLanguageServerReady() || !serviceProvider.Options.ApiToken.IsValid()) return;
-                        var sastSettings = await this.serviceProvider.LanguageClientManager.InvokeGetSastEnabled(SnykVSPackage.Instance.DisposalToken);
-
-                        bool snykCodeEnabled = sastSettings != null ? sastSettings.SnykCodeEnabled : false;
-
-                        this.UpdateSnykCodeEnablementSettings(sastSettings);
-
-                        if (snykCodeEnabled)
-                        {
-                            this.snykCodeEnableTimer.Stop();
-                        }
-                        else if (currentRequestAttempt < MaxSastRequestAttempts)
-                        {
-                            currentRequestAttempt++;
-
-                            this.snykCodeEnableTimer.Interval = TwoSecondsDelay * currentRequestAttempt;
-                        }
-                        else
-                        {
-                            this.snykCodeEnableTimer.Stop();
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        this.HandleSastError(e);
-                    }
-                });
-
-                this.snykCodeEnableTimer.Start();
-            }
-            catch (Exception e)
-            {
-                this.HandleSastError(e);
-            }
-        }
-
-        private void HandleSastError(Exception e)
-        {
-            this.snykCodeEnableTimer.Stop();
-
-            NotificationService.Instance.ShowErrorInfoBar(e.Message);
-
-            this.codeSecurityEnabledCheckBox.Enabled = false;
-
-            this.snykCodeDisabledInfoLabel.Visible = false;
-            this.snykCodeSettingsLinkLabel.Visible = false;
-            this.checkAgainLinkLabel.Visible = false;
-        }
-
-        private void UpdateSnykCodeEnablementSettings(SastSettings sastSettings)
-        {
-            var snykCodeEnabled = sastSettings?.SnykCodeEnabled ?? false;
-
-            if (!snykCodeEnabled)
-            {
-                this.snykCodeDisabledInfoLabel.Text = "Snyk Code is disabled by your organisation\'s configuration:";
-            }
-
-            this.codeSecurityEnabledCheckBox.Enabled = snykCodeEnabled;
-            this.snykCodeDisabledInfoLabel.Visible = !snykCodeEnabled;
-            this.snykCodeSettingsLinkLabel.Visible = !snykCodeEnabled;
-            this.checkAgainLinkLabel.Visible = !snykCodeEnabled;
+            // SAST enablement check removed - checkbox is always toggleable
         }
 
         private void UpdateViewFromOptions()
         {
             this.ossEnabledCheckBox.Checked = OptionsMemento.OssEnabled;
             this.iacEnabledCheckbox.Checked = OptionsMemento.IacEnabled;
+            this.codeSecurityEnabledCheckBox.Checked = OptionsMemento.SnykCodeSecurityEnabled;
 
             if (cbDelta.DataSource == null)
             {
@@ -177,11 +92,6 @@ namespace Snyk.VisualStudio.Extension.Settings
 
         private void SnykCodeSettingsLinkLabel_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
             => Process.Start(OptionsMemento.SnykCodeSettingsUrl);
-
-        private void CheckAgainLinkLabel_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-        {
-            this.StartSastEnablementCheckLoop();
-        }
 
     }
 }